refactor(Gender): Change format of Gender Enum
