+++ src/main/scala/com/particeep/api/models/enums/Gender.scala
@@ -1,14 +1,12 @@
 package com.particeep.api.models.enums
 
-object Gender {
+sealed trait Gender extends Product with Enum
 
-  sealed abstract class Gender extends Product with Enum
-  case object MAN              extends Gender
-  case object WOMAN            extends Gender
+object Gender extends EnumHelper[Gender] {
+  case object MAN   extends Gender
+  case object WOMAN extends Gender
 
-  object Gender extends EnumHelper[Gender] {
-    def values: Set[Gender] = Set(MAN, WOMAN)
+  def values: Set[Gender] = Set(MAN, WOMAN)
 
-    implicit def string2gender(value: String): Gender = get(value).getOrElse(MAN)
-  }
+  implicit def string2gender(value: String): Gender = get(value).getOrElse(MAN)
 }
