+++ src/main/scala/com/particeep/api/models/user/Relative.scala
@@ -7,7 +7,7 @@ import play.api.libs.json.{ JsObject, OFormat, Writes }
 import java.time.OffsetDateTime
 
 import com.particeep.api.core.Formatter
-import com.particeep.api.models.enums.Gender.Gender
+import com.particeep.api.models.enums.Gender
 import com.particeep.api.models.enums.RelativeType.RelativeType
 
 case class Relative(
